+++ components/restreamer/src/file_manager.rs
@@ -5,17 +5,16 @@ use std::{
     path::PathBuf,
 };
 
-use derive_more::{Deref, Display, Into};
+use derive_more::{Deref, Display, From, Into};
 use ephyr_log::{tracing, Instrument};
 use juniper::{GraphQLEnum, GraphQLObject, GraphQLScalar, ScalarValue};
 use serde::{Deserialize, Serialize};
 use tap::prelude::*;
 
 use crate::{
+    api::google_drive::{responses::FileInfo as DriveFileInfo, GoogleDriveApi},
     cli::Opts,
-    display_panic,
-    file_manager::api_response::{get_gdrive_result, ErrorResponse},
-    spec,
+    display_panic, spec,
     state::{InputEndpointKind, InputSrc, State, Status},
     stream_probe::stream_probe,
     stream_statistics::StreamStatistics,
@@ -23,16 +22,9 @@ use crate::{
 use chrono::Utc;
 use ephyr_log::tracing::instrument;
 use futures::{FutureExt, TryFutureExt};
-use reqwest::{Response, StatusCode};
 use std::{
     borrow::BorrowMut, ffi::OsString, fs::DirEntry, panic::AssertUnwindSafe,
-    result::Result::Err,
 };
-
-const GDRIVE_PUBLIC_PARAMS: &str = "supportsAllDrives=True\
-&supportsTeamDrives=True\
-&includeItemsFromAllDrives=True\
-&includeTeamDriveItems=True";
 
 /// Commands for handling operations on files
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -60,6 +52,7 @@ pub enum FileCommand {
     Eq,
     Hash,
     Into,
+    From,
     PartialEq,
     Default,
     Serialize,
@@ -236,19 +229,12 @@ impl FileManager {
         api_key: &'a str,
         state: &'a State,
     ) -> Result<(), String> {
-        let response = reqwest::get(
-            format!(
-                "https://www.googleapis.com/drive/v3/files/{file_id}?
-                fields=name&key={api_key}&{GDRIVE_PUBLIC_PARAMS}"
-            )
-            .as_str(),
-        )
-        .await;
-
-        let filename =
-            get_gdrive_result::<api_response::FileNameResponse>(response)
-                .await?
-                .name;
+        let filename = GoogleDriveApi::new(api_key)
+            .files()
+            .get_file_info(file_id.as_str())
+            .await
+            .map_err(|e| format!("{e}"))?
+            .name;
 
         state
             .files
@@ -258,8 +244,7 @@ impl FileManager {
             .map_or_else(
                 || {
                     tracing::error!(
-                        "Could not find file \
-                             with the provided id: {}",
+                        "Could not find file with the provided id: {}",
                         file_id
                     );
                     Err("Could not find the provided file ID".to_string())
@@ -286,13 +271,6 @@ impl FileManager {
                     .clone()
                     .ok_or("No API key provided")?;
 
-                let client = reqwest::ClientBuilder::new()
-                    .connection_verbose(false)
-                    .build()
-                    .map_err(|err| {
-                        format!("Could not create a reqwest Client: {err}")
-                    })?;
-
                 // Get file name from the API
                 if file_name.is_none() {
                     Self::update_file_info(&file_id, &api_key, &state)
@@ -318,87 +296,45 @@ impl FileManager {
                 }
 
                 // Download the file contents
-                if let Ok(mut response) = client
-                    .get(
-                        format!(
-                            "https://www.googleapis.com/drive/v3/files/\
-                            {file_id}?alt=media&key={api_key}\
-                            &{GDRIVE_PUBLIC_PARAMS}"
-                        )
-                        .as_str(),
-                    )
-                    .send()
+                let mut response = GoogleDriveApi::new(&api_key)
+                    .files()
+                    .get_file_response(&file_id)
                     .await
-                {
-                    let status = response.status();
-                    if status != StatusCode::OK {
-                        if status == 403 {
-                            return Err(response
-                                .json::<ErrorResponse>()
-                                .await
-                                .map(|r| {
-                                    format!(
-                                        "Http response: {} {}",
-                                        r.error.code, r.error.message
-                                    )
-                                })
-                                .map_err(|e| {
-                                    format!(
-                                        "Status: {status}. Unknown error {e}"
-                                    )
-                                })?);
-                        }
+                    .map_err(|e| format!("{e}"))?;
 
-                        let error_response = response.text().await;
-                        return Err(format!(
-                            "Can't download file. Http response: {} {}",
-                            status,
-                            error_response.unwrap_or_default(),
-                        ));
-                    }
-
-                    let total = response.content_length();
-                    // Create FileInfo Download state and set the state
-                    // to Downloading
-                    state
-                        .files
-                        .lock_mut()
-                        .iter_mut()
-                        .find(|file| file.file_id == file_id)
-                        .ok_or_else(|| {
-                            "Could not find file with the \
-                             provided file ID"
-                                .to_string()
-                        })?
-                        .pipe_borrow_mut(|val| {
-                            val.download_state = Some(DownloadState {
-                                max_progress: NetworkByteSize(total.unwrap()),
-                                current_progress: NetworkByteSize(0),
-                            });
-                            val.state = FileState::Downloading;
+                let total = response.content_length();
+                // Create DriveFileInfo Download state and set the state
+                // to Downloading
+                state
+                    .files
+                    .lock_mut()
+                    .iter_mut()
+                    .find(|file| file.file_id == file_id)
+                    .ok_or_else(|| {
+                        "Could not find file with the provided file ID"
+                            .to_string()
+                    })?
+                    .pipe_borrow_mut(|val| {
+                        val.download_state = Some(DownloadState {
+                            max_progress: NetworkByteSize(total.unwrap()),
+                            current_progress: NetworkByteSize(0),
                         });
+                        val.state = FileState::Downloading;
+                    });
 
-                    Self::download_and_write_bytes(
-                        &file_id,
-                        &root_dir,
-                        response.borrow_mut(),
-                        &state,
-                    )
-                    .await?;
+                Self::download_and_write_bytes(
+                    &file_id,
+                    &root_dir,
+                    response.borrow_mut(),
+                    &state,
+                )
+                .await?;
 
-                    Ok(response.status().as_u16())
-                } else {
-                    Err("Could not send download request for the file"
-                        .to_string())
-                }
+                Ok(response.status().as_u16())
             }
             .await
             .map_err(|err| {
-                tracing::error!(
-                    "Could not download file {}: {}",
-                    &file_id,
-                    err
-                );
+                tracing::error!("Could not download file {file_id}: {err}",);
                 state
                     .files
                     .lock_mut()
@@ -424,7 +360,7 @@ impl FileManager {
     async fn download_and_write_bytes(
         file_id: &FileId,
         root_dir: &str,
-        response: &mut Response,
+        response: &mut reqwest::Response,
         state: &State,
     ) -> Result<(), String> {
         // Try opening the target file where the downloaded
@@ -451,7 +387,7 @@ impl FileManager {
             }
 
             current.0 += bytes.len() as u64;
-            // Update download progress in the FileInfo,
+            // Update download progress in the DriveFileInfo,
             // but only each 400ms
             if Utc::now()
                 .signed_duration_since(last_update)
@@ -502,6 +438,7 @@ impl FileManager {
                 file.download_state.as_mut().unwrap().current_progress =
                     current;
                 file.state = FileState::Local;
+                file.error = None;
                 update_stream_info(
                     file_id.clone(),
                     file_path.clone(),
@@ -584,10 +521,10 @@ pub struct LocalFileInfo {
     pub download_state: Option<DownloadState>,
 }
 
-impl From<api_response::ExtendedFileInfoResponse> for LocalFileInfo {
-    fn from(file_response: api_response::ExtendedFileInfoResponse) -> Self {
+impl From<DriveFileInfo> for LocalFileInfo {
+    fn from(file_response: DriveFileInfo) -> Self {
         LocalFileInfo {
-            file_id: FileId(file_response.id),
+            file_id: FileId::from(file_response.id),
             name: Some(file_response.name),
             state: FileState::Pending,
             download_state: None,
@@ -612,12 +549,10 @@ pub struct PlaylistFileInfo {
     pub was_played: bool,
 }
 
-impl From<api_response::ExtendedFileInfoResponse>
-    for spec::v1::PlaylistFileInfo
-{
-    fn from(file_response: api_response::ExtendedFileInfoResponse) -> Self {
+impl From<DriveFileInfo> for spec::v1::PlaylistFileInfo {
+    fn from(file_response: DriveFileInfo) -> Self {
         spec::v1::PlaylistFileInfo {
-            file_id: FileId(file_response.id),
+            file_id: FileId::from(file_response.id),
             name: file_response.name,
         }
     }
@@ -695,6 +630,31 @@ impl NetworkByteSize {
     }
 }
 
+/// Retrieves file information from Google Drive given an API key and file ID.
+///
+/// # Errors
+///
+/// Returns an error message as a `String` in the following cases:
+///
+/// - If there is an issue with the Google Drive API request.
+/// - If the file is not a video.
+pub async fn get_video_file_from_gdrive(
+    api_key: &str,
+    file_id: &str,
+) -> Result<spec::v1::PlaylistFileInfo, String> {
+    let file_info = GoogleDriveApi::new(api_key)
+        .files()
+        .get_file_info(file_id)
+        .await
+        .map_err(|e| format!("{e}"))?;
+
+    if file_info.is_video() {
+        Ok(file_info.into())
+    } else {
+        Err("This is not video file".to_string())
+    }
+}
+
 /// Retrieves list of video files from a Google drive folder
 ///
 /// # Errors
@@ -704,126 +664,11 @@ pub async fn get_video_list_from_gdrive_folder(
     api_key: &str,
     folder_id: &str,
 ) -> Result<Vec<spec::v1::PlaylistFileInfo>, String> {
-    let mut response =
-        api_response::FileListResponse::retrieve_dir_content_from_api(
-            api_key, folder_id,
-        )
+    let response = GoogleDriveApi::new(api_key)
+        .files()
+        .get_dir_videos(folder_id)
+        .map_err(|e| format!("{e}"))
         .await?;
-    response.filter_only_video_files();
-    Ok(response
-        .files
-        .drain(..)
-        .map(|x| spec::v1::PlaylistFileInfo {
-            file_id: FileId(x.id),
-            name: x.name,
-        })
-        .collect())
-}
 
-pub(crate) mod api_response {
-    use crate::file_manager::GDRIVE_PUBLIC_PARAMS;
-    use reqwest::{Response, StatusCode};
-    use serde::Deserialize;
-
-    /// Used to deserialize Google API call for the file details
-    #[derive(Deserialize)]
-    pub(crate) struct FileNameResponse {
-        /// Name of the file
-        pub(crate) name: String,
-    }
-
-    #[derive(Deserialize, Debug)]
-    pub(crate) struct ExtendedFileInfoResponse {
-        pub(crate) id: String,
-        pub(crate) name: String,
-        #[serde(alias = "mimeType")]
-        pub(crate) mime_type: String,
-    }
-
-    impl ExtendedFileInfoResponse {
-        #[allow(dead_code)]
-        pub(crate) fn is_dir(&self) -> bool {
-            self.mime_type == "application/vnd.google-apps.folder"
-        }
-
-        pub(crate) fn is_video(&self) -> bool {
-            self.mime_type.starts_with("video")
-        }
-    }
-
-    #[derive(Deserialize)]
-    pub(crate) struct FileListResponse {
-        // TODO fix this
-        // pub(crate) kind: String,
-        // pub(crate) incomplete_search: bool,
-        pub(crate) files: Vec<ExtendedFileInfoResponse>,
-    }
-
-    impl FileListResponse {
-        pub(crate) async fn retrieve_dir_content_from_api(
-            api_key: &str,
-            dir_id: &str,
-        ) -> Result<Self, String> {
-            let response = reqwest::get(
-                format!(
-                    "https://www.googleapis.com/drive/v3/files?\
-                     key={api_key}&q='{dir_id}'%20in%20parents&\
-                     fields=files/id,files/name,files/mimeType&\
-                     {GDRIVE_PUBLIC_PARAMS}",
-                )
-                .as_str(),
-            )
-            .await;
-            get_gdrive_result::<Self>(response).await
-        }
-
-        pub(crate) fn filter_only_video_files(&mut self) {
-            self.files.retain(ExtendedFileInfoResponse::is_video);
-        }
-    }
-
-    ///
-    #[derive(Deserialize)]
-    pub(crate) struct ErrorResponse {
-        pub(crate) error: ErrorMessage,
-    }
-
-    ///
-    #[derive(Deserialize)]
-    pub(crate) struct ErrorMessage {
-        pub(crate) code: u16,
-        pub(crate) message: String,
-    }
-
-    pub(crate) async fn get_gdrive_result<T: for<'de> Deserialize<'de>>(
-        response: reqwest::Result<Response>,
-    ) -> Result<T, String> {
-        let error_parsing = |e| format!("Error parsing JSON result: {e}");
-
-        match response {
-            Err(e) => Err(format!("No valid response from the API: {e}")),
-            Ok(r) => {
-                let status = r.status();
-                if status == StatusCode::OK {
-                    return r.json::<T>().await.map_err(error_parsing);
-                } else if status == 403 {
-                    return Err(r
-                        .json::<ErrorResponse>()
-                        .await
-                        .map(|x| {
-                            format!(
-                                "Http response: {} {}",
-                                x.error.code, x.error.message
-                            )
-                        })
-                        .map_err(error_parsing)?);
-                }
-
-                Err(r
-                    .text()
-                    .await
-                    .map_err(|e| format!("Unknown error: {e}"))?)
-            }
-        }
-    }
+    Ok(response.into_iter().map(Into::into).collect())
 }