+++ components/restreamer/src/teamspeak.rs
@@ -265,7 +265,7 @@ impl fmt::Debug for Input {
             .field("audio", &"Arc<Mutex<AudioHandler>>")
             .field("conn", &self.conn)
             .field("is_conn_unrecoverable", &self.is_conn_unrecoverable)
-            .finish()
+            .finish_non_exhaustive()
     }
 }
 
@@ -429,8 +429,11 @@ impl Future for AudioCapture {
         loop {
             let StreamItem::Audio(audio_packet) =
                 ready!(Pin::new(&mut self.conn.events()).poll_next(cx))
-                .ok_or_else(|| E::UnexpectedFinish)?
-                .map_err(E::ConnectionFailed)? else { continue };
+                    .ok_or_else(|| E::UnexpectedFinish)?
+                    .map_err(E::ConnectionFailed)?
+            else {
+                continue;
+            };
 
             let member_id = match audio_packet.data().data() {
                 AudioData::S2C { from, .. }